requirements.txt
nose2
nose-cov
